pyproject.toml
[project]
name = "soccer-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asttokens==3.0.0",
    "async-timeout==4.0.3",
    "beautifulsoup4>=4.13.4",
    "bottleneck==1.5.0",
    "cachetools==5.5.2",
    "certifi==2025.6.15",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "comm==0.2.2",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "debugpy==1.8.14",
    "decorator==5.2.1",
    "docstring-parser==0.16",
    "exceptiongroup==1.3.0",
    "executing==2.2.0",
    "fastapi>=0.116.1",
    "filetype==1.2.0",
    "fonttools==4.58.4",
    "google-ai-generativelanguage==0.6.18",
    "google-api-core==2.25.1",
    "google-auth==2.40.3",
    "google-cloud-aiplatform==1.103.0",
    "google-cloud-bigquery==3.34.0",
    "google-cloud-core==2.4.3",
    "google-cloud-resource-manager==1.14.2",
    "google-cloud-storage==2.19.0",
    "google-crc32c==1.7.1",
    "google-genai==1.25.0",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.3",
    "grpc-google-iam-v1==0.14.2",
    "grpcio==1.73.1",
    "grpcio-status==1.73.1",
    "h11==0.16.0",
    "html5lib>=1.1",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "idna==3.10",
    "ipykernel==6.29.5",
    "ipython==8.37.0",
    "jedi==0.19.2",
    "joblib==1.5.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "kiwisolver==1.4.8",
    "langchain==0.3.26",
    "langchain-core==0.3.68",
    "langchain-google-genai==2.1.8",
    "langchain-google-vertexai==2.0.27",
    "langchain-mcp-adapters==0.1.7",
    "langchain-text-splitters==0.3.8",
    "langgraph==0.4.8",
    "langgraph-checkpoint==2.1.0",
    "langgraph-cli>=0.3.5",
    "langgraph-prebuilt==0.2.3",
    "langgraph-sdk==0.1.70",
    "langsmith==0.3.45",
    "lxml==5.4.0",
    "matplotlib==3.10.3",
    "matplotlib-inline==0.1.7",
    "mcp==1.9.4",
    "nest-asyncio==1.6.0",
    "numexpr==2.11.0",
    "numpy==2.2.6",
    "orjson==3.10.18",
    "ormsgpack==1.10.0",
    "packaging==24.2",
    "pandas==2.3.0",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pillow==11.2.1",
    "platformdirs==4.3.8",
    "prompt-toolkit==3.0.51",
    "proto-plus==1.26.1",
    "protobuf==6.31.1",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyarrow==19.0.1",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pygments==2.19.2",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "pyzmq==27.0.0",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "rsa==4.9.1",
    "scikit-learn==1.7.0",
    "scipy==1.15.3",
    "seaborn==0.13.2",
    "shapely==2.1.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "sse-starlette==2.3.6",
    "stack-data==0.6.3",
    "starlette==0.47.1",
    "tenacity==8.5.0",
    "threadpoolctl==3.6.0",
    "tornado==6.5.1",
    "traitlets==5.14.3",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.34.3",
    "validators==0.35.0",
    "wcwidth==0.2.13",
    "websockets==15.0.1",
    "xxhash==3.5.0",
    "zstandard==0.23.0",
]
